Revert "Merge pull request #2 from geeksam/hash-with-dubious-semantics"
This reverts commit 6ae0bdb458acde2f85a3422808b2f32bdd3bec21, reversing
changes made to 1cd8437e1af56997a6c91f4f515cc6439a215cc5.

The HashWithDubiousSemantics was a good idea, but the execution makes
it so the params are no longer recognized as a Hash.  This causes
several problems with client apps.  We'll probably want to take an
approach that's similar to ActiveSupport's HashWithIndifferentAccess.